+++ apps/openmw/mwclass/misc.cpp
@@ -90,13 +90,13 @@ namespace MWClass
                 if (Settings::game().mRebalanceSoulGemValues)
                 {
                     // use the 'soul gem value rebalance' formula from the Morrowind Code Patch
-                    float soulValue = 0.0001 * pow(soul, 3) + 2 * soul;
+                    double soulValue = 0.0001 * std::pow(soul, 3) + 2 * soul;
 
                     // for Azura's star add the unfilled value
                     if (ptr.getCellRef().getRefId() == "Misc_SoulGem_Azura")
-                        value += soulValue;
+                        value += static_cast<int>(soulValue);
                     else
-                        value = soulValue;
+                        value = static_cast<int>(soulValue);
                 }
                 else
                     value *= soul;
